% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_getauthortitles}
\alias{bhl_getauthortitles}
\title{Return a list of titles associated with a given BHL author identifier.}
\usage{
bhl_getauthortitles(creatorid, as = "table", key = NULL, ...)
}
\arguments{
\item{creatorid}{BHL identifier for a particular author (numeric)}

\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Unless the identifier  for a particular BHL author record is known in
   advance, this method should be used in combination	with the AuthorSearch
   method.
}
\examples{
\dontrun{
bhl_getauthortitles(1970)
bhl_getauthortitles(1970, as='json')
bhl_getauthortitles(1970, as='xml')
bhl_getauthortitles(1970, as='list')
}
}

