% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_getcollections}
\alias{bhl_getcollections}
\title{Get a list of collections which are used to group titles and items. A single
collection may contain either titles or items, but not both.}
\usage{
bhl_getcollections(as = "table", key = NULL, ...)
}
\arguments{
\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get a list of collections which are used to group titles and items. A single
collection may contain either titles or items, but not both.
}
\examples{
\dontrun{
bhl_getcollections()
bhl_getcollections(as = 'list')
bhl_getcollections(as = 'json')
bhl_getcollections(as = 'xml')
}
}

