% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getinstitutions.R
\name{bhl_getinstitutions}
\alias{bhl_getinstitutions}
\title{GetInstitutions data}
\usage{
bhl_getinstitutions(as = "table", key = NULL, ...)
}
\arguments{
\item{as}{(character) Return a list ("list"), json ("json"), xml ("xml"), or parsed table
("table", default). Note that \code{as="table"} can give different data format back
depending on the function - for example, sometimes a data.frame and sometimes a
character vector.}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
GetInstitutions data
}
\details{
Get a list of institutions which have contributed to BHL in some
way. These institutions may fill roles such as contributors, rights
holders, and scanning institutions.
}
\examples{
\dontrun{
bhl_getinstitutions()
}
}

