% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getpagenames.R
\name{bhl_getpagenames}
\alias{bhl_getpagenames}
\title{Return a list of names that appear on a page.}
\usage{
bhl_getpagenames(...)
}
\description{
Return a list of names that appear on a page.
}
\keyword{internal}
