% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getpageocrtext.R
\name{bhl_getpageocrtext}
\alias{bhl_getpageocrtext}
\title{Return the OCR-generated text of a page.}
\usage{
bhl_getpageocrtext(...)
}
\description{
Return the OCR-generated text of a page.
}
\keyword{internal}
