% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getpartbibtex.R
\name{bhl_getpartbibtex}
\alias{bhl_getpartbibtex}
\title{Return a citation for a part, using the BibTeX format.}
\usage{
bhl_getpartbibtex(...)
}
\description{
Return a citation for a part, using the BibTeX format.
}
\keyword{internal}
