% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bhl_getpartbibtex}
\alias{bhl_getpartbibtex}
\title{Return a citation for a part, using the BibTeX format.}
\usage{
bhl_getpartbibtex(partid, format = "json", output = "list", key = NULL,
  callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{partid}{The identifier of an individual part (article, chapter, etc) (numeric)}
}
\description{
Return a citation for a part, using the BibTeX format.
}
\examples{
\dontrun{
bhl_getpartbibtex(1000)
bhl_getpartbibtex(1000, output='raw')
bhl_getpartbibtex(1000, format='xml', output='raw')
bhl_getpartbibtex(1000, format='xml', output='parsed')
}
}

