% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_getpartbyidentifier}
\alias{bhl_getpartbyidentifier}
\title{Return a list of the identifiers of all unpublished items.}
\usage{
bhl_getpartbyidentifier(type = NULL, value = NULL, format = "json",
  output = "list", key = NULL, ...)
}
\arguments{
\item{type}{The type of identifier (doi, oclc, issn, isbn, lccn, ddc, nal, nlm, coden)}

\item{value}{The identifier value}

\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of the identifiers of all unpublished items.
}
\examples{
\dontrun{
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12')
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12', output='raw')
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12', format='xml', output='raw')
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12', format='xml', output='parsed')
}
}

