% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getpartnames.R
\name{bhl_getpartnames}
\alias{bhl_getpartnames}
\title{Return a list of scientific names associated with a part}
\usage{
bhl_getpartnames(...)
}
\description{
Return a list of scientific names associated with a part
}
\keyword{internal}
