% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getsubjectparts.R
\name{bhl_getsubjectparts}
\alias{bhl_getsubjectparts}
\title{Return a list of parts (articles, chapters, etc) associated with a subject}
\usage{
bhl_getsubjectparts(...)
}
\description{
Return a list of parts (articles, chapters, etc) associated with a subject
}
\keyword{internal}
