% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bhl_getsubjectparts.R
\name{bhl_getsubjectparts}
\alias{bhl_getsubjectparts}
\title{Return a list of parts (articles, chapters, etc) associated with a subject.}
\usage{
bhl_getsubjectparts(subject = NULL, as = "table", key = NULL, ...)
}
\arguments{
\item{subject}{the full or partial subject for which to search (character)}

\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Note: haven't seen examples for this function that work yet...
}
\examples{
\dontrun{
bhl_getsubjectparts('frogs')
bhl_getsubjectparts('diptera', 'xml')
bhl_getsubjectparts('diptera', 'json')
}
}

