% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getsubjecttitles.R
\name{bhl_getsubjecttitles}
\alias{bhl_getsubjecttitles}
\title{Return a list of titles associated with a subject}
\usage{
bhl_getsubjecttitles(...)
}
\description{
Return a list of titles associated with a subject
}
\keyword{internal}
