% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_gettitlebibtex.R
\name{bhl_gettitlebibTex}
\alias{bhl_gettitlebibTex}
\title{Return a citation for a title, using the BibTeX format}
\usage{
bhl_gettitlebibTex(...)
}
\description{
Return a citation for a title, using the BibTeX format
}
\keyword{internal}
