% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_gettitlebyidentifier.R
\name{bhl_gettitlebyidentifier}
\alias{bhl_gettitlebyidentifier}
\title{Find and return metadata about a title or titles that match a specific
identifier}
\usage{
bhl_gettitlebyidentifier(...)
}
\description{
Find and return metadata about a title or titles that match a specific
identifier
}
\keyword{internal}
