% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_gettitleendNote}
\alias{bhl_gettitleendNote}
\title{Return a citation for a title, using the EndNote format.}
\usage{
bhl_gettitleendNote(titleid = NA, key = NULL, ...)
}
\arguments{
\item{titleid}{the identifier of an individual title (numeric)}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}
}
\description{
Return a citation for a title, using the EndNote format.
}
\examples{
\dontrun{
bhl_gettitleendNote(1726)
}
}

