% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_gettitleitems.R
\name{bhl_gettitleitems}
\alias{bhl_gettitleitems}
\title{Return a list of a title's items (books)}
\usage{
bhl_gettitleitems(...)
}
\description{
Return a list of a title's items (books)
}
\keyword{internal}
