% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getunpublishedparts.R
\name{bhl_getunpublishedparts}
\alias{bhl_getunpublishedparts}
\title{Return a list of the identifiers of all unpublished parts
(articles, chapters, etc)}
\usage{
bhl_getunpublishedparts(...)
}
\description{
Return a list of the identifiers of all unpublished parts
(articles, chapters, etc)
}
\keyword{internal}
