% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getunpublishedtitles.R
\name{bhl_getunpublishedtitles}
\alias{bhl_getunpublishedtitles}
\title{Return a list of the identifiers of all unpublished titles}
\usage{
bhl_getunpublishedtitles(...)
}
\description{
Return a list of the identifiers of all unpublished titles
}
\keyword{internal}
