% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_namecount.R
\name{bhl_namecount}
\alias{bhl_namecount}
\title{Return the number of unique names found on pages in BHL}
\usage{
bhl_namecount(...)
}
\description{
Return the number of unique names found on pages in BHL
}
\keyword{internal}
