% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpages.R
\name{bhl_pages}
\alias{bhl_pages}
\title{Get many OCR-generated pages given a single item id}
\usage{
bhl_pages(itemid, key = NULL, ...)
}
\arguments{
\item{itemid}{the item id (character). Required}

\item{key}{Your BHL API key, either enter, or loads from your
\code{.Renviron} as \code{BHL_KEY} or from \code{.Rprofile} as
\code{bhl_key}}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Get many OCR-generated pages given a single item id
}
\examples{
\dontrun{
x <- bhl_publicationsearch('Selborne', year=1825, collection=4,
  language='eng')
bhl_pages(itemid = x$ItemID[1])
}
}
