% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_titlesearchsimple.R
\name{bhl_titlesearchsimple}
\alias{bhl_titlesearchsimple}
\title{Perform a simple title search}
\usage{
bhl_titlesearchsimple(...)
}
\description{
Perform a simple title search
}
\keyword{internal}
