\name{query2message}
\alias{query2message}
\title{Internal function to convert API query to message printed in the console.}
\usage{
  query2message(x)
}
\arguments{
  \item{x}{List of named parameters to be used in API
  call.}
}
\description{
  Internal function to convert API query to message printed
  in the console.
}
\examples{
\dontrun{
message(query2message(args))
}
}

