% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbhl-package.R
\name{rbhl-defunct}
\alias{rbhl-defunct}
\title{Defunct functions in rbhl}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
\item \code{bhl_getpartendnote()}: BHL removed this API method.
\item \code{bhl_gettitleendnote()}: BHL removed this API method.
}
}
