% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbhl-package.R
\docType{package}
\name{rbhl-package}
\alias{rbhl-package}
\alias{rbhl}
\title{R interface to the Biodiversity Heritage Library API.}
\description{
R interface to the Biodiversity Heritage Library API.
}
\details{
You need an API key to use the Biodiversity Heritage Library API. Get your
BHL API key at \url{http://www.biodiversitylibrary.org/getapikey.aspx}.
Put your API key in your .Rprofile file using e.g.,
`options(BioHerLibKey = "YOURBHLAPIKEY")`, and the functions within
this package will be able to use your API key without you having to enter
it every time you run a search.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
