% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbhl-package.R
\docType{data}
\name{rbhlmethods}
\alias{rbhlmethods}
\title{Data.frame of all the BHL API methods from the BHL website.}
\description{
Data.frame of all the BHL API methods from the BHL website.
}
\keyword{datasets}

