% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhl_authorsearch}
\alias{bhl_authorsearch}
\title{Return a list of authors that match (fully or partially) the specified
   search string.}
\usage{
bhl_authorsearch(name = NULL, format = "json", output = "list",
  key = NULL, callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{name}{full or partial name of the author for which to search
(last name listed first, i.e. 'Darwin, Charles') (character)}
}
\description{
The namessearched are those contained in MARC 100a, 110a, 111a, 700a,
   710a, and 711a library records.
}
\examples{
\dontrun{
bhl_authorsearch(name='dimmock')
bhl_authorsearch(name='Jones')
}
}

