% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_getauthortitles}
\alias{bhl_getauthortitles}
\title{Return a list of titles associated with a given BHL author identifier.}
\usage{
bhl_getauthortitles(creatorid = NA, format = "json", output = "list",
  key = NULL, ...)
}
\arguments{
\item{creatorid}{BHL identifier for a particular author (numeric)}

\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Unless the identifier  for a particular BHL author record is known in
   advance, this method should be used in combination	with the AuthorSearch
   method.
}
\examples{
\dontrun{
bhl_getauthortitles(1970)
bhl_getauthortitles(1970, output='raw')
bhl_getauthortitles(1970, output='raw', format='xml')
}
}

