% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bhl_getcollections}
\alias{bhl_getcollections}
\title{Get a list of collections which are used to group titles and items. A single
   collection may contain either titles or items, but not both.}
\usage{
bhl_getcollections(format = "json", output = "list", key = NULL,
  callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}
}
\description{
Get a list of collections which are used to group titles and items. A single
   collection may contain either titles or items, but not both.
}
\examples{
\dontrun{
bhl_getcollections()
bhl_getcollections(out = 'raw')
}
}

