% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bhl_getitemmetadata}
\alias{bhl_getitemmetadata}
\title{Return metadata about an item.}
\usage{
bhl_getitemmetadata(itemid = NULL, pages = TRUE, ocr = FALSE,
  parts = FALSE, format = "json", output = "list", key = NULL,
  callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{itemid}{item id (character)}

\item{pages}{return the items pages (TRUE/FALSE)}

\item{ocr}{(logical) TRUE to return the ocr for the item's pages. Setting this
to TRUE apparently doesn't return any actual ocr text, but leaving parameter
here for now.}

\item{parts}{(logical) TRUE to return the item's parts. Setting this
to TRUE apparently doesn't return any parts text, but leaving parameter
here for now.}
}
\description{
You may choose to include a list of the item's pages.
}
\examples{
\dontrun{
bhl_getitemmetadata('16800', TRUE)
bhl_getitemmetadata('16800', TRUE, 'xml', 'parsed')
bhl_getitemmetadata('16800', TRUE, 'json', 'raw')
bhl_getitemmetadata('16800', TRUE, 'xml', 'raw')


bhl_getitemmetadata(20419, pages=FALSE, parts=TRUE)
}
}

