% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhl_getitempages}
\alias{bhl_getitempages}
\title{Return a list of an item's pages.}
\usage{
bhl_getitempages(itemid, ocr = FALSE, format = "json", output = "list",
  key = NULL, callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{itemid}{the item id (character)}

\item{ocr}{return ocr text of the page (TRUE/FALSE)}
}
\description{
Return a list of an item's pages.
}
\examples{
\dontrun{
bhl_getitempages('16800')
bhl_getitempages('16800', 'xml', 'raw')
bhl_getitempages('16800', 'xml', 'parsed')

# Return ocr text
bhl_getitempages('16800', TRUE)
}
}

