% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bhl_getlanguages.R
\name{bhl_getlanguages}
\alias{bhl_getlanguages}
\title{Get a list of languages in which materials in BHL have been written.}
\usage{
bhl_getlanguages(as = "table", key = NULL, ...)
}
\arguments{
\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get a list of languages in which materials in BHL have been written.
}
\examples{
\dontrun{
bhl_getlanguages()
bhl_getlanguages('json')
bhl_getlanguages('xml')
bhl_getlanguages('list')
}
}

