% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bhl_getpagemetadata.R
\name{bhl_getpagemetadata}
\alias{bhl_getpagemetadata}
\title{Return metadata about a page.}
\usage{
bhl_getpagemetadata(page = NULL, ocr = FALSE, names = FALSE,
  as = "table", key = NULL, ...)
}
\arguments{
\item{page}{page number to get}

\item{ocr}{return ocr text of the page (TRUE/FALSE)}

\item{names}{return the names that appear on the page (TRUE/FALSE)}

\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
You may choose to include the OCR text and a list of names found on the page.
}
\examples{
\dontrun{
bhl_getpagemetadata(page=1328690, ocr=TRUE)
bhl_getpagemetadata(page=1328690, ocr=TRUE, as='json')
bhl_getpagemetadata(page=1328690, ocr=TRUE, as='xml')
bhl_getpagemetadata(page=1328690, ocr=TRUE, as='list')
}
}

