% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhl_getpagenames}
\alias{bhl_getpagenames}
\title{Return a list of names that appear on a page.}
\usage{
bhl_getpagenames(page = NULL, format = "json", output = "list",
  key = NULL, callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{page}{page number to get}
}
\description{
Return a list of names that appear on a page.
}
\examples{
\dontrun{
bhl_getpagenames('1328690')
bhl_getpagenames('1328690', 'xml', 'raw')
bhl_getpagenames('1328690', 'xml', 'parsed')
}
}

