% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_getpageocrtext}
\alias{bhl_getpageocrtext}
\title{Return the OCR-generated text of a page.}
\usage{
bhl_getpageocrtext(page = NULL, format = "json", output = "list",
  key = NULL, ...)
}
\arguments{
\item{page}{page number to get}

\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return the OCR-generated text of a page.
}
\examples{
\dontrun{
bhl_getpageocrtext(1328690, 'json')
bhl_getpageocrtext(1328690, 'xml', 'raw')
bhl_getpageocrtext(1328690, 'xml', 'parsed')
}
}

