% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bhl_getpartbyidentifier}
\alias{bhl_getpartbyidentifier}
\title{Return a list of the identifiers of all unpublished items.}
\usage{
bhl_getpartbyidentifier(type = NULL, value = NULL, format = "json",
  output = "list", key = NULL, callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{type}{The type of identifier (doi, oclc, issn, isbn, lccn, ddc, nal, nlm, coden)}

\item{value}{The identifier value}
}
\description{
Return a list of the identifiers of all unpublished items.
}
\examples{
\dontrun{
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12')
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12', output='raw')
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12', format='xml', output='raw')
bhl_getpartbyidentifier('doi', '10.4039/Ent38406-12', format='xml', output='parsed')
}
}

