% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_getpartmetadata}
\alias{bhl_getpartmetadata}
\title{Return a list of an item's pages.}
\usage{
bhl_getpartmetadata(partid, format = "json", output = "list", key = NULL,
  ...)
}
\arguments{
\item{partid}{The identifier of an individual part (article, chapter, etc) (numeric)}

\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of an item's pages.
}
\examples{
\dontrun{
bhl_getpartmetadata(10409)
bhl_getpartmetadata(10409, 'xml', 'raw')
bhl_getpartmetadata(10409, 'xml', 'parsed')
}
}

