% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhl_getpartmetadata}
\alias{bhl_getpartmetadata}
\title{Return a list of an item's pages.}
\usage{
bhl_getpartmetadata(partid, format = "json", output = "list", key = NULL,
  callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{partid}{The identifier of an individual part (article, chapter, etc) (numeric)}
}
\description{
Return a list of an item's pages.
}
\examples{
\dontrun{
bhl_getpartmetadata(10409)
bhl_getpartmetadata(10409, 'xml', 'raw')
bhl_getpartmetadata(10409, 'xml', 'parsed')
}
}

