% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getpartmetadata.R
\name{bhl_getpartmetadata}
\alias{bhl_getpartmetadata}
\title{Return a list of an item's pages.}
\usage{
bhl_getpartmetadata(partid, key = NULL, ...)
}
\arguments{
\item{partid}{The identifier of an individual part (article, chapter, etc) (numeric)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of an item's pages.
}
\examples{
\dontrun{
bhl_getpartmetadata(10409)
}
}
