% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhl_getpartnames}
\alias{bhl_getpartnames}
\title{Return a list of scientific names associated with a part.}
\usage{
bhl_getpartnames(partid, format = "json", output = "list", key = NULL,
  callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{partid}{The identifier of an individual part (article, chapter, etc) (character)}
}
\description{
Return a list of scientific names associated with a part.
}
\examples{
\dontrun{
bhl_getpartnames(7443)
}
}

