% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bhl_getsubjecttitles.R
\name{bhl_getsubjecttitles}
\alias{bhl_getsubjecttitles}
\title{Return a list of titles associated with a subject.}
\usage{
bhl_getsubjecttitles(subject = NULL, as = "table", key = NULL, ...)
}
\arguments{
\item{subject}{the full or partial subject for which to search (character)}

\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of titles associated with a subject.
}
\examples{
\dontrun{
bhl_getsubjecttitles('diptera')
bhl_getsubjecttitles('diptera', 'xml')
bhl_getsubjecttitles('diptera', 'json')
}
}

