% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_gettitleendnote.R
\name{bhl_gettitleendNote}
\alias{bhl_gettitleendNote}
\title{Return a citation for a title, using the EndNote format.}
\usage{
bhl_gettitleendNote(titleid = NA, key = NULL, ...)
}
\arguments{
\item{titleid}{the identifier of an individual title (numeric)}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a citation for a title, using the EndNote format.
}
\examples{
\dontrun{
bhl_gettitleendNote(1726)
}
}

