% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhl_getunpublisheditems}
\alias{bhl_getunpublisheditems}
\title{Return a list of the identifiers of all unpublished items.}
\usage{
bhl_getunpublisheditems(format = "json", output = "list", key = NULL,
  callopts = list())
}
\arguments{
\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{your BHL API key, either enter, or loads from .Rprofile}

\item{callopts}{Call options passed on to httr::GET.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}
}
\description{
Return a list of the identifiers of all unpublished items.
}
\examples{
\dontrun{
bhl_getunpublisheditems()
bhl_getunpublisheditems(output='raw')
bhl_getunpublisheditems(format='xml', output='raw')
bhl_getunpublisheditems(format='xml', output='parsed')
}
}

