% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhl_getunpublisheditems.R
\name{bhl_getunpublisheditems}
\alias{bhl_getunpublisheditems}
\title{Return a list of the identifiers of all unpublished items.}
\usage{
bhl_getunpublisheditems(as = "table", key = NULL, ...)
}
\arguments{
\item{as}{(character) Return a list ("list"), json ("json"), xml ("xml"), or parsed table
("table", default). Note that \code{as="table"} can give different data format back
depending on the function - for example, sometimes a data.frame and sometimes a
character vector.}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of the identifiers of all unpublished items.
}
\examples{
\dontrun{
bhl_getunpublisheditems()
bhl_getunpublisheditems('xml')
bhl_getunpublisheditems('json')
}
}
