% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bhl_getunpublishedtitles.R
\name{bhl_getunpublishedtitles}
\alias{bhl_getunpublishedtitles}
\title{Return a list of the identifiers of all unpublished titles.}
\usage{
bhl_getunpublishedtitles(as = "table", key = NULL, ...)
}
\arguments{
\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of the identifiers of all unpublished titles.
}
\examples{
\dontrun{
bhl_getunpublishedtitles()
bhl_getunpublishedtitles('json')
bhl_getunpublishedtitles('xml')
}
}

