% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bhl_openurl.R
\name{bhl_openurl}
\alias{bhl_openurl}
\title{Not sure how this differs from their other API...}
\usage{
bhl_openurl(genre = NULL, title = NULL, aufirst = NULL, aulast = NULL,
  date = NULL, spage = NULL, issue = NULL, version = 0.1, as = "list",
  key = NULL, ...)
}
\arguments{
\item{genre}{Book genre}

\item{title}{Book title}

\item{aufirst}{First author}

\item{aulast}{Last author}

\item{date}{Date of publication}

\item{spage}{Start page}

\item{issue}{Issue number}

\item{version}{One of 0.1 or 1.0}

\item{as}{(character) Return a list ("list"), json ("json"), xml ("xml"), or parsed table
("table", default). Note that \code{as="table"} can give different data format back
depending on the function - for example, sometimes a data.frame and sometimes a
character vector.}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Not sure how this differs from their other API...
}
\examples{
\dontrun{
bhl_openurl(
	genre="book",
	title="Manual+of+North+American+Diptera",
	aufirst="Samuel Wendell",
	aulast="Williston",
	date=1908,
	spage=16)

bhl_openurl(genre="book", title="Manual+of+North+American+Diptera",
   aufirst="Samuel Wendell", aulast="Williston", date=1908, spage=16)

bhl_openurl(genre="book", title="Manual+of+North+American+Diptera",
   aufirst="Samuel Wendell", aulast="Williston", date=1908, spage=16, as='xml')
}
}

