% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_openurl}
\alias{bhl_openurl}
\title{Not sure how this differs from their other API...}
\usage{
bhl_openurl(genre = NULL, title = NULL, aufirst = NULL, aulast = NULL,
  date = NULL, spage = NULL, issue = NULL, version = 0.1,
  format = "json", key = NULL, output = "list", ...)
}
\arguments{
\item{genre}{Book genre}

\item{title}{Book title}

\item{aufirst}{First author}

\item{aulast}{Last author}

\item{date}{Date of publication}

\item{spage}{Start page}

\item{issue}{Issue number}

\item{version}{One of 0.1 or 1.0}

\item{format}{either XML ('xml') or JSON ('json') format. If output='list',
format is forced to equal 'json'}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{output}{Return a list, raw json or xml, or parsed data (character).
Default: 'list'}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Not sure how this differs from their other API...
}
\examples{
\dontrun{
bhl_openurl(
	genre="book",
	title="Manual+of+North+American+Diptera",
	aufirst="Samuel Wendell",
	aulast="Williston",
	date=1908,
	spage=16)

bhl_openurl(genre="book", title="Manual+of+North+American+Diptera",
   aufirst="Samuel Wendell", aulast="Williston", date=1908, spage=16, output='raw')

bhl_openurl(genre="book", title="Manual+of+North+American+Diptera",
   aufirst="Samuel Wendell", aulast="Williston", date=1908, spage=16, output='parsed')

bhl_openurl(genre="book", title="Manual+of+North+American+Diptera",
   aufirst="Samuel Wendell", aulast="Williston", date=1908, spage=16,
   format='xml', output='parsed')
}
}

