% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhl_subjectsearch}
\alias{bhl_subjectsearch}
\title{Return a list of subjects that match (fully or partially) the specified
   search string.}
\usage{
bhl_subjectsearch(subject, as = "table", key = NULL, ...)
}
\arguments{
\item{subject}{the full or partial subject for which to search (character)}

\item{as}{(character) Return a list, json xml, or parsed table (default)}

\item{key}{Your BHL API key, either enter, or loads from your \code{.Renviron} as \code{BHL_KEY}
or from \code{.Rprofile} as \code{bhl_key}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Return a list of subjects that match (fully or partially) the specified
   search string.
}
\examples{
\dontrun{
bhl_subjectsearch('diptera')
bhl_subjectsearch('diptera', "json")
}
}

