% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getpages.R
\name{getpages}
\alias{getpages}
\title{Get many OCR-generated pages given a single item id}
\usage{
getpages(itemid = NULL, key = NULL, ...)
}
\arguments{
\item{itemid}{the item id (character)}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get many OCR-generated pages given a single item id
}
\examples{
\dontrun{
books <- bhl_booksearch(title='Selborne', lname='White', volume=2, edition='new', year=1825,
   collectionid=4, language='eng')
bhl_getpages(itemid=16800)
}
}

