% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbhl-package.R
\docType{package}
\name{rbhl-package}
\alias{rbhl-package}
\alias{rbhl}
\title{rbhl}
\description{
R interface to the Biodiversity Heritage Library API.

You need an API key to use the Biodiversity Heritage Library API. Get your
BHL API key at \url{http://www.biodiversitylibrary.org/getapikey.aspx}.
Put your API key either in your .Renviron file (or similar) with the name
\code{BHL_KEY}, or in your .Rprofile file with the name \code{bhl_key}, and the
we'll use your API key without you having to enter
it every time you run a search.

See \link{rbhl-defunct} for defunct functions.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
